requirements.txt
# Beverly Knits ERP - Consolidated Requirements
# Python 3.10+ required

# Core Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.3.0
werkzeug>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
xlsxwriter>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0

# Machine Learning & Forecasting
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.0
statsmodels>=0.14.0
river>=0.18.0  # Online learning
joblib>=1.3.0
scipy>=1.7.0

# Dependency Injection
dependency-injector>=4.41.0

# Async & Background Tasks
celery>=5.3.0
kombu>=5.3.0

# API & Utilities
requests>=2.31.0
aiohttp>=3.8.5          # Async HTTP client for eFab API
tenacity>=8.2.0         # Retry logic for API resilience
cryptography>=41.0.0    # For secure credential storage
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0

# Production Server
gunicorn>=21.2.0
uvicorn>=0.23.0

# Monitoring
prometheus-client>=0.17.0

# Testing (for development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional - Large packages
# tensorflow>=2.15.0  # Only if LSTM needed (1GB+)
# matplotlib>=3.7.0   # Only for visualization
# seaborn>=0.12.0     # Only for visualization
# plotly>=5.15.0      # Only for interactive plots